Set rpcTlsPort to have a default value of 443
When compiling from source, the default value of NULL in settings->rpcTlsPort causes the false error:

Network error: Couldn't resolve host name

This is due to the fact that during login, the URL is built according to:

(ui.c line 245)

BarUiPianoCall (BarApp_t * const app, const PianoRequestType_t PIANO_REQUEST_LOGIN,
                void * const data, PianoReturn_t * const pRet, CURLcode * const wRet)

In this function, we receive the initial (unformed) request in the form of a pointer to data, which comes from main.c, specifically line 67 in function BarMainLoginUser

        reqData.user = app->settings.username;
        reqData.password = app->settings.password;
        reqData.step = 0;

This gets moved into a new allocation of the PianoRequest_t data type (line 253 ui.c), a pointer of which is passed to PianoRequest() along with PIANO_REQUEST_LOGIN:

    req->urlPath = PIANO_RPC_PATH "method=auth.partnerLogin" # comes out to /services/json?method=auth.partnerLogin

    req->secure is also set to true in this method.

Back in BarUiPianoCall, we now call BarPianoHttpRequest.

Here's where the problem is. To this point, settings->rpcTlsPort=NULL.

When we now build the URL:

        int ret = snprintf (url, sizeof (url), "%s://%s:%s%s",
                req->secure ? "https" : "http",
                settings->rpcHost,
                req->secure ? settings->rpcTlsPort : "80",
                req->urlPath);

It comes out to be:

    https://tunder.pandora.com:NULL/services/json?method=auth.partnerLogin

Which leads the CURL request to fail, and exit with an error.

Setting rpcTlsPort to 443 in settings.c will fix this error; current fix is to have "rpc_tls_port = 443" in your $XDG_CONFIG_HOME/pianobar/config file

Hopefully this pull request gets added to prevent the frustrations of future source builders!